# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "TestGyver API",
    "description": "API pour la plateforme de test multi-environnements TestGyver - Really, are you serious?",
    "version": "1.0.0",
    "contact": {
      "name": "TestGyver Support"
    }
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
    }
  },
...
  "paths": {
...
    "/campains": {
      "get": {
        "tags": ["Campagnes"],
        "summary": "Liste des campagnes",
        "security": [{"Bearer": []}],
        "responses": {
          "200": {
            "description": "Liste des campagnes",
            "schema": {
              "type": "array",
              "items": {"$ref": "#/definitions/Campain"}
            }
          }
        }
      },
      "post": {
        "tags": ["Campagnes"],
        "summary": "Créer une campagne",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "body",
            "name": "campain",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CampainCreate"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Campagne créée",
            "schema": {
              "$ref": "#/definitions/Campain"
            }
          },
          "400": {"description": "Données invalides"}
        }
      }
    },
    "/campains/import": {
      "post": {
        "tags": ["Campagnes"],
        "summary": "Importer une campagne",
        "description": "Importe une campagne depuis un fichier JSON exporté",
        "security": [{"Bearer": []}],
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "in": "formData",
            "name": "file",
            "type": "file",
            "required": true,
            "description": "Fichier JSON de la campagne"
          },
          {
            "in": "formData",
            "name": "name",
            "type": "string",
            "required": false,
            "description": "Nom de la campagne importée (optionnel)"
          }
        ],
        "responses": {
          "201": {
            "description": "Campagne importée",
            "schema": {
              "type": "object",
              "properties": {
                "message": {"type": "string"},
                "campain_id": {"type": "string"}
              }
            }
          },
          "400": {"description": "Fichier invalide ou manquant"}
        }
      }
    },
    "/campains/{campain_id}": {
      "get": {
        "tags": ["Campagnes"],
        "summary": "Détails d'une campagne",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "path",
            "name": "campain_id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Détails de la campagne",
            "schema": {
              "$ref": "#/definitions/Campain"
            }
          },
          "404": {"description": "Campagne non trouvée"}
        }
      },
      "put": {
        "tags": ["Campagnes"],
        "summary": "Mettre à jour une campagne",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "path",
            "name": "campain_id",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "campain",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CampainUpdate"
            }
          }
        ],
        "responses": {
          "200": {"description": "Campagne mise à jour"},
          "404": {"description": "Campagne non trouvée"}
        }
      },
      "delete": {
        "tags": ["Campagnes"],
        "summary": "Supprimer une campagne",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "path",
            "name": "campain_id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {"description": "Campagne supprimée"},
          "404": {"description": "Campagne non trouvée"}
        }
      }
    },
    "/campains/{campain_id}/files": {
      "get": {
        "tags": ["Campagnes"],
        "summary": "Liste des fichiers d'une campagne",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "path",
            "name": "campain_id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des fichiers",
            "schema": {
              "type": "array",
              "items": {"$ref": "#/definitions/CampainFile"}
            }
          },
          "404": {"description": "Campagne non trouvée"}
        }
      },
      "post": {
        "tags": ["Campagnes"],
        "summary": "Upload un fichier dans une campagne",
        "security": [{"Bearer": []}],
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "in": "path",
            "name": "campain_id",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "file",
            "type": "file",
            "required": true,
            "description": "Fichier à uploader"
          },
          {
            "in": "formData",
            "name": "rename",
            "type": "string",
            "required": false,
            "description": "Nouveau nom du fichier (optionnel)"
          }
        ],
        "responses": {
          "201": {"description": "Fichier uploadé"},
          "400": {"description": "Fichier manquant"},
          "404": {"description": "Campagne non trouvée"}
        }
      }
    },
    "/campains/{campain_id}/files/{filename}": {
      "get": {
        "tags": ["Campagnes"],
        "summary": "Télécharger un fichier de campagne",
        "security": [{"Bearer": []}],
        "produces": ["application/octet-stream"],
        "parameters": [
          {
            "in": "path",
            "name": "campain_id",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "filename",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {"description": "Fichier téléchargé"},
          "404": {"description": "Fichier non trouvé"}
        }
      },
      "delete": {
        "tags": ["Campagnes"],
        "summary": "Supprimer un fichier de campagne",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "path",
            "name": "campain_id",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "filename",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {"description": "Fichier supprimé"},
          "404": {"description": "Fichier non trouvé"}
        }
      }
    },
    "/campains/{campain_id}/export": {
      "get": {
        "tags": ["Campagnes"],
        "summary": "Exporter une campagne",
        "description": "Exporte la campagne au format JSON avec option d'inclure les fichiers",
        "security": [{"Bearer": []}],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "campain_id",
            "type": "string",
            "required": true,
            "description": "ID de la campagne"
          },
          {
            "in": "query",
            "name": "include_files",
            "type": "boolean",
            "description": "Inclure les fichiers du répertoire de travail (encodés en base64)",
            "default": false
          }
        ],
        "responses": {
          "200": {
            "description": "Fichier JSON de la campagne",
            "schema": {
              "type": "file"
            }
          },
          "404": {"description": "Campagne non trouvée"}
        }
      }
    },
...
    "/variables": {
      "get": {
        "tags": ["Variables"],
        "summary": "Liste des variables",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "query",
            "name": "grouped",
            "type": "boolean",
            "description": "Grouper par filière"
          },
          {
            "in": "query",
            "name": "filiere",
            "type": "string",
            "description": "Filtrer par filière"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des variables",
            "schema": {
              "type": "array",
              "items": {"$ref": "#/definitions/Variable"}
            }
          }
        }
      },
      "post": {
        "tags": ["Variables"],
        "summary": "Créer une variable",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "body",
            "name": "variable",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VariableCreate"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Variable créée",
            "schema": {
              "$ref": "#/definitions/Variable"
            }
          },
          "400": {"description": "Données invalides"}
        }
      }
    },
    "/variables/{variable_id}": {
      "get": {
        "tags": ["Variables"],
        "summary": "Détails d'une variable",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "path",
            "name": "variable_id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Détails de la variable",
            "schema": {
              "$ref": "#/definitions/Variable"
            }
          },
          "404": {"description": "Variable non trouvée"}
        }
      },
      "put": {
        "tags": ["Variables"],
        "summary": "Mettre à jour une variable",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "path",
            "name": "variable_id",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "variable",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VariableUpdate"
            }
          }
        ],
        "responses": {
          "200": {"description": "Variable mise à jour"},
          "404": {"description": "Variable non trouvée"}
        }
      },
      "delete": {
        "tags": ["Variables"],
        "summary": "Supprimer une variable",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "path",
            "name": "variable_id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {"description": "Variable supprimée"},
          "404": {"description": "Variable non trouvée"}
        }
      }
    },
...
    "/deleted/campains": {
      "get": {
        "tags": ["Éléments supprimés"],
        "summary": "Liste des campagnes supprimées logiquement",
        "description": "Récupère toutes les campagnes supprimées logiquement (admin uniquement)",
        "security": [{"Bearer": []}],
        "responses": {
          "200": {
            "description": "Liste des campagnes supprimées",
            "schema": {
              "type": "object",
              "properties": {
                "campains": {
                  "type": "array",
                  "items": {"$ref": "#/definitions/Campain"}
                }
              }
            }
          },
          "401": {"description": "Non authentifié"},
          "403": {"description": "Accès interdit (admin requis)"}
        }
      }
    },
...
  },
  "definitions": {
...
    "Campain": {
      "type": "object",
      "properties": {
        "_id": {"type": "string"},
        "userCreated": {"type": "string"},
        "name": {"type": "string"},
        "dateCreated": {"type": "string", "format": "date-time"},
        "description": {"type": "string"}
      }
    },
    "CampainCreate": {
      "type": "object",
      "required": ["name"],
      "properties": {
        "name": {"type": "string"},
        "description": {"type": "string"}
      }
    },
    "CampainUpdate": {
      "type": "object",
      "properties": {
        "name": {"type": "string"},
        "description": {"type": "string"}
      }
    },
    "CampainFile": {
      "type": "object",
      "properties": {
        "name": {"type": "string"},
        "size": {"type": "integer", "description": "Taille en octets"},
        "modified": {"type": "string", "format": "date-time"}
      }
    },
...
    "Variable": {
      "type": "object",
      "properties": {
        "_id": {"type": "string"},
        "key": {"type": "string"},
        "value": {},
        "filiere": {"type": "string"},
        "description": {"type": "string"},
        "isRoot": {"type": "boolean", "default": false}
      }
    },
    "VariableCreate": {
      "type": "object",
      "required": ["key", "value"],
      "properties": {
        "key": {"type": "string"},
        "value": {},
        "filiere": {"type": "string"},
        "description": {"type": "string"},
        "isRoot": {"type": "boolean"}
      }
    },
    "VariableUpdate": {
      "type": "object",
      "properties": {
        "key": {"type": "string"},
        "value": {},
        "filiere": {"type": "string"},
        "description": {"type": "string"},
        "isRoot": {"type": "boolean"}
      }
    },
...
  }
}